+++ cub3d.c
@@ -6,7 +6,7 @@
 /*   By: thisai <thisai@student.42tokyo.jp>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/13 16:23:13 by thisai            #+#    #+#             */
-/*   Updated: 2021/02/22 12:32:26 by thisai           ###   ########.fr       */
+/*   Updated: 2021/02/22 12:58:21 by thisai           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -241,267 +241,6 @@ double	c3_distance_squared(t_c3_vector *p1, t_c3_vector *p2)
 	dx = p1->x - p2->x;
 	dy = p1->y - p2->y;
 	return (dx * dx + dy * dy);
-}
-
-int		c3_check_wall(t_c3_state *stat, t_c3_vector *hit)
-{
-	if (hit->x < 0 || hit->x >= stat->map.width
-		|| hit->y < 0 || hit->y >= stat->map.height)
-		return (1);
-	if (c3_query_map(stat, hit->x, hit->y) == C3_MAP_SYMBOL_WALL)
-		return (1);
-	return (0);
-}
-
-int		c3_check_sprite(
-	t_c3_state *stat, t_c3_vector *center,
-	t_c3_vector *pos, t_c3_hit_result *result)
-{
-	if (c3_query_map(stat, center->x, center->y) == C3_MAP_SYMBOL_SPRITE)
-	{
-		result->type = C3_MAP_SYMBOL_SPRITE;
-		result->distance_sqared = c3_distance_squared(pos, center);
-		return (1);
-	}
-	return (0);
-}
-
-double	c3_dot(t_c3_vector *origin, t_c3_vector *a, t_c3_vector *b)
-{
-	return (a->x - origin->x) * (b->x - origin->x)
-		+ (a->y - origin->y) * (b->y - origin->y);
-}
-
-int		c3_add_sprite(
-	t_c3_vector *hit, t_c3_vector *hit_cell, t_c3_vector *pos,
-	t_c3_hit_result *result)
-{
-	double		dot;
-	t_c3_vector	ad;
-	double		c;
-	double		offset;
-
-	dot = c3_dot(pos, hit_cell, hit);
-	ad.x = hit->x - pos->x;
-	ad.y = hit->y - pos->y;
-	c = result[1].distance_sqared / dot;
-	hit->x = pos->x + ad.x * c;
-	hit->y = pos->y + ad.y * c;
-
-	offset = sqrt(c3_distance_squared(hit, hit_cell));
-	if (offset <= 0.5)
-	{
-		t_c3_vector	ab = {hit_cell->x - pos->x, hit_cell->y - pos->y};
-		double	cross = ab.x * ad.y - ad.x * ab.y;
-
-		if (cross > 0)
-			offset = - offset;
-
-		result[1].position = *hit;
-		result[1].offset = offset;
-		return (1);
-	}
-	return (0);
-}
-
-t_c3_vector	c3_get_horizontal_hit_for_index(
-	double theta, int index, t_c3_vector *pos)
-{
-	t_c3_vector	hit;
-
-	if (theta >= 0 && theta < M_PI)
-	{
-		hit.y = floor(pos->y) + index;
-		hit.x = pos->x + (hit.y - pos->y) / tan(theta);
-	}
-	else
-	{
-		hit.y = floor(pos->y) - index + 1;
-		hit.x = pos->x + (hit.y - pos->y) / tan(theta);
-	}
-	return (hit);
-}
-
-t_c3_vector	c3_get_horizontal_hit_cell_for_index(
-	double theta, t_c3_vector *hit)
-{
-	t_c3_vector	hit_cell;
-
-	if (theta >= 0 && theta < M_PI)
-	{
-		hit_cell.x = floor(hit->x) + 0.5;
-		hit_cell.y = floor(hit->y) + 0.5;
-	}
-	else
-	{
-		hit_cell.x = floor(hit->x) + 0.5;
-		hit_cell.y = floor(hit->y) - 1 + 0.5;
-	}
-	return (hit_cell);
-}
-
-int		c3_get_horizontal_hit(
-	t_c3_state *stat, t_c3_vector *pos,
-	double theta, t_c3_hit_result *result)
-{
-	t_c3_vector	hit;
-	t_c3_vector	hit_cell;
-	int			facing_north;
-	int			index;
-	int			hit_sprites;
-
-	hit_sprites = 0;
-	index = 1;
-	while (1)
-	{
-		facing_north = !(theta >= 0 && theta < M_PI);
-		hit = c3_get_horizontal_hit_for_index(theta, index, pos);
-		hit_cell = c3_get_horizontal_hit_cell_for_index(theta, &hit);
-		if (c3_check_wall(stat, &hit_cell))
-			break ;
-		if (hit_sprites < C3_MAX_COLLINEAR_SPRITES
-			&& c3_check_sprite(stat, &hit_cell, pos, &result[hit_sprites + 1]))
-			if(c3_add_sprite(&hit, &hit_cell, pos, result + hit_sprites))
-				hit_sprites ++;
-		index++;
-	}
-	result->position = hit;
-	result->distance_sqared = c3_distance_squared(pos, &hit);
-	result->type = facing_north ? C3_OBJTYPE_WALL_N : C3_OBJTYPE_WALL_S;
-	return (hit_sprites);
-}
-
-t_c3_vector	c3_get_vertical_hit_for_index(
-	double theta, int index, t_c3_vector *pos)
-{
-	t_c3_vector	hit;
-
-	if (theta < M_PI_2 || theta >= 3 * M_PI_2)
-	{
-		hit.x = floor(pos->x) + index;
-		hit.y = pos->y + (hit.x - pos->x) * tan(theta);
-	}
-	else
-	{
-		hit.x = floor(pos->x) - index + 1;
-		hit.y = pos->y + (hit.x - pos->x) * tan(theta);
-	}
-	return (hit);
-}
-
-t_c3_vector	c3_get_vertical_hit_cell_for_index(
-	double theta, t_c3_vector *hit)
-{
-	t_c3_vector	hit_cell;
-
-	if (theta < M_PI_2 || theta >= 3 * M_PI_2)
-	{
-		hit_cell.x = floor(hit->x) + 0.5;
-		hit_cell.y = floor(hit->y) + 0.5;
-	}
-	else
-	{
-		hit_cell.x = floor(hit->x) - 1 + 0.5;
-		hit_cell.y = floor(hit->y) + 0.5;
-	}
-	return (hit_cell);
-}
-
-int		c3_get_vertical_hit(
-	t_c3_state *stat, t_c3_vector *pos,
-	double theta, t_c3_hit_result *result)
-{
-	t_c3_vector	hit;
-	t_c3_vector	hit_cell;
-	int			facing_east;
-	int			index;
-	int			hit_sprites;
-
-	hit_sprites = 0;
-	index = 1;
-	while (1)
-	{
-		facing_east = (theta < M_PI_2 || theta >= 3 * M_PI_2);
-		hit = c3_get_vertical_hit_for_index(theta, index, pos);
-		hit_cell = c3_get_vertical_hit_cell_for_index(theta, &hit);
-		if (c3_check_wall(stat, &hit_cell))
-			break ;
-		if (hit_sprites < C3_MAX_COLLINEAR_SPRITES
-			&& c3_check_sprite(stat, &hit_cell, pos, &result[hit_sprites + 1]))
-			if(c3_add_sprite(&hit, &hit_cell, pos, result + hit_sprites))
-				hit_sprites ++;
-		index++;
-	}
-	result->position = hit;
-	result->distance_sqared = c3_distance_squared(pos, &hit);
-	result->type = facing_east ? C3_OBJTYPE_WALL_E : C3_OBJTYPE_WALL_W;
-	return (hit_sprites);
-}
-
-int		c3_cast_ray(
-	t_c3_state *stat, t_c3_vector *pos, double theta, t_c3_hit_result *out)
-{
-	t_c3_hit_result	hori_hits[1 + C3_MAX_COLLINEAR_SPRITES];
-	t_c3_hit_result	vert_hits[1 + C3_MAX_COLLINEAR_SPRITES];
-	int				hori_sprites;
-	int				vert_sprites;
-	int				hori_index;
-	int				vert_index;
-	int				sprites;
-	double			wall_distance;
-
-	hori_sprites = tan(theta) != 0.0 ? c3_get_horizontal_hit(stat, pos, theta, hori_hits) : 0;
-	vert_sprites = c3_get_vertical_hit(stat, pos, theta, vert_hits);
-
-	if (tan(theta) != 0.0 &&
-		hori_hits[0].distance_sqared < vert_hits[0].distance_sqared)
-	{
-		*out = hori_hits[0];
-		wall_distance = hori_hits[0].distance_sqared;
-	}
-	else
-	{
-		*out = vert_hits[0];
-		wall_distance = vert_hits[0].distance_sqared;
-	}
-
-	hori_index = 0;
-	vert_index = 0;
-	sprites = 0;
-	while (sprites < C3_MAX_COLLINEAR_SPRITES
-		   && (hori_index < hori_sprites || vert_index < vert_sprites))
-	{
-		if (hori_index >= hori_sprites && vert_index < vert_sprites)
-		{
-			if (vert_hits[vert_index + 1].distance_sqared >= wall_distance)
-				vert_index++;
-			else
-				out[sprites++ + 1] = vert_hits[vert_index++ + 1];
-		}
-		else if (hori_index < hori_sprites && vert_index >= vert_sprites)
-		{
-			if (hori_hits[hori_index + 1].distance_sqared >= wall_distance)
-				hori_index++;
-			else
-				out[sprites++ + 1] = hori_hits[hori_index++ + 1];
-		}
-		else if (hori_hits[hori_index + 1].distance_sqared
-				 < vert_hits[vert_index + 1].distance_sqared)
-		{
-			if (hori_hits[vert_index + 1].distance_sqared >= wall_distance)
-				hori_index++;
-			else
-				out[sprites++ + 1] = hori_hits[hori_index++ + 1];
-		}
-		else
-		{
-			if (vert_hits[vert_index + 1].distance_sqared >= wall_distance)
-				vert_index++;
-			else
-				out[sprites++ + 1] = vert_hits[vert_index++ + 1];
-		}
-	}
-	return (sprites);
 }
 
 void	c3_draw_rays_on_map(t_c3_state *stat)
@@ -744,112 +483,6 @@ void	c3_draw_walls(t_c3_state *stat)
 void	c3_render_scene(t_c3_state *stat)
 {
 	c3_draw_walls(stat);
-}
-
-void	c3_draw(t_c3_state *stat)
-{
-	c3_render_scene(stat);
-	/* c3_draw_map(stat); */
-
-	mlx_put_image_to_window(stat->mlx, stat->window, stat->img, 0, 0);
-
-	/* c3_draw_rays_on_map(stat); */
-
-	/* mlx_string_put( */
-	/* 	stat->mlx, stat->window, 10, 10, mlx_get_color_value(stat->mlx, 0xffffff), "CUB3D"); */
-}
-
-void	c3_scan(t_c3_state *stat)
-{
-	int	x;
-	int	half_res;
-	int	hit_sprites;
-
-	half_res = stat->renderer.resolution_x / 2;
-	x = -half_res;
-	while (x < half_res)
-	{
-		double a =
-			x * (stat->renderer.plane_distance * tan(stat->renderer.fov))
-			/ half_res;
-		double theta = atan(a / stat->renderer.plane_distance);
-		stat->renderer.rays[x + half_res].angle = theta;
-
-		double angle = theta + stat->player.direction;
-		if (angle >= M_PI * 2)
-			angle -= M_PI * 2;
-		if (angle < 0)
-			angle += M_PI * 2;
-		hit_sprites = c3_cast_ray(
-			stat, &stat->player.position, angle,
-			stat->renderer.rays[x + half_res].hits);
-		stat->renderer.rays[x + half_res].hit_sprite_count = hit_sprites;
-		x++;
-	}
-}
-
-void	c3_update(t_c3_state *stat)
-{
-	if (stat->keystate.left)
-	{
-		stat->player.direction -= stat->player.rotation_speed;
-		while (stat->player.direction < 0)
-			stat->player.direction += M_PI * 2;
-	}
-	else if (stat->keystate.right)
-	{
-		stat->player.direction += stat->player.rotation_speed;
-		while (stat->player.direction >= M_PI * 2)
-			stat->player.direction -= M_PI * 2;
-	}
-
-	if (stat->keystate.w || stat->keystate.s)
-	{
-		double delta_x = stat->player.walk_speed * cos(stat->player.direction);
-		double delta_y = stat->player.walk_speed * sin(stat->player.direction);
-		double new_x = stat->player.position.x + delta_x * (stat->keystate.w ? 1 : -1);
-		double new_y = stat->player.position.y + delta_y * (stat->keystate.w ? 1 : -1);
-		if (new_x >= 0 && new_x < stat->map.width
-			&& new_y >= 0 && new_y < stat->map.height)
-		{
-			stat->player.position.x = new_x;
-			stat->player.position.y = new_y;
-		}
-	}
-
-	if (stat->keystate.a || stat->keystate.d)
-	{
-		double delta_x = stat->player.walk_speed * sin(stat->player.direction);
-		double delta_y = stat->player.walk_speed * -cos(stat->player.direction);
-		double new_x = stat->player.position.x + delta_x * (stat->keystate.a ? 1 : -1);
-		double new_y = stat->player.position.y + delta_y * (stat->keystate.a ? 1 : -1);
-		if (new_x >= 0 && new_x < stat->map.width
-			&& new_y >= 0 && new_y < stat->map.height)
-		{
-			stat->player.position.x = new_x;
-			stat->player.position.y = new_y;
-		}
-	}
-
-	c3_scan(stat);
-}
-
-int		c3_loop_hook(void *param)
-{
-	t_c3_state	*stat;
-	/* static int	current_frame; */
-
-	stat = (t_c3_state*)param;
-
-	/* if (current_frame++ > 1000) */
-	/* { */
-	/* 	c3_terminate(stat); */
-	/* 	exit(0); */
-	/* } */
-
-	c3_update(stat);
-	c3_draw(stat);
-	return (1);
 }
 
 void	c3_texture_cache_init(t_c3_texture_cache *cache, t_c3_scene *scene)
